Add ability to have notes on print out
This patch as an overlay canvas to the generated PDF print version to
render the notes added to the slides. To keep the current default behavior,
this feature must be enabled by setting the new configuration option "printNotes"
to `true`.

Unfortunatly the relative positioning between the drawings and the slide content
is not perfect since th rendering of the slide content slightly differs in the
pdf-print mode.

This patch adresses issue #97